Break Bullet out to its own file
